requirements.txt
# ECH0-PRIME Core Dependencies
# Copyright (c) 2025 Joshua Hendricks Cole (DBA: Corporation of Light). All Rights Reserved. PATENT PENDING.

# Core Scientific Computing
numpy>=1.24.0
pandas

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Quantum Computing
qiskit>=0.45.0
qiskit-aer>=0.13.0
qiskit-ibm-provider>=0.11.0

# Image Processing (Vision Bridge)
Pillow>=10.0.0

# HTTP/API Communication
requests>=2.31.0
flask>=3.0.0
flask-cors>=4.0.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Audio & Voice Processing
SpeechRecognition>=3.10.0
PyAudio>=0.2.13
elevenlabs>=0.2.27

# LLM & AI Services
python-dotenv>=1.0.0
datasets>=2.14.0
huggingface_hub
transformers>=4.30.0
pinecone
sentence-transformers
openai>=1.3.0
anthropic>=0.7.0

# Utility
psutil>=5.9.0
pydantic>=2.5.0
websockets>=11.0
feedparser>=6.0.10
psycopg2-binary>=2.9.9
redis>=5.0.0
opencv-python pymupdf langchain-text-splitters

# Vector Databases
faiss-cpu>=1.7.4
# qdrant-client>=1.6.0  # Optional: uncomment if using Qdrant
# weaviate-client>=3.24.0  # Optional: uncomment if using Weaviate

# Causal Inference
dowhy>=0.11.0
causal-learn>=0.1.3
networkx>=3.0
scipy>=1.10.0

# Probabilistic Programming
pyro-ppl>=1.8.6

# AI Benchmarks
datasets>=2.14.0
evaluate>=0.4.0
transformers>=4.30.0
accelerate>=0.20.0
peft>=0.4.0
bitsandbytes>=0.41.0
scipy>=1.10.0
scikit-learn>=1.3.0
tqdm>=4.65.0
openai>=1.3.0
anthropic>=0.7.0
